refactor: deduplicate fixtures in errata tests
fake_koji and fake_errata_tool were set up identically across
several tests. Adding them to conftest.py allows removing this
duplicate code.
